requirements.txt
setuptools~=69.0.3
python-dotenv~=1.0.0